requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
aiosqlite
greenlet
jinja2
pydantic-settings
passlib

#fastapi-users[sqlalchemy]
#fastapi-users[password]
#fastapi-users[redis]
#fastapi-users

fastapi-admin

aiofiles
asyncpg

sentry-sdk[fastapi]
pillow

ruff
mypy



fastapi-csrf-protect

pydub
audioop-lts # fix problem pydub in python 3.13
ffmpeg-python

openai

nltk

#psycopg2
psycopg2-binary

# USERS
fastapi[all]
python-jose[cryptography]
#passlib[bcrypt]
passlib
bcrypt==4.0.1

pydantic
email-validator

sentry-sdk[fastapi]

requests

slowapi
